+++ CHANGELOG.md
@@ -1,8 +1,10 @@
+## unreleased
+* Add `networkTransactionId` to `CreditCardVerification`
+
 ## 5.2.0
 * Add `retrieval_reference_number` to `Transaction`
 * Fix class alias issue where Exceptions could not be thrown for missing libraries or older PHP versions.
 * Fix issue where `proxyHost`, `proxyPort`, `proxyType`, `proxyUser`, `proxyPassword`, and `sslVersion` were not being set on Gateway or Configuration objects.
-* Add `networkTransactionId` to `CreditCardVerification`
 
 ## 5.1.0
 * Add `threeDSecurePassThru` parameters to `Customer.create`, `PaymentMethod.create`, `CreditCard.create`, `Customer.update`, `PaymentMethod.update` and `CreditCard.update`